+++ tests/test-files/naming-convention/ok.ts
@@ -1,0 +1,21 @@
+class Moo {
+    private _x?: number;
+
+    private someFunction() {
+        return this._x;
+    }
+}
+
+new Moo();
+
+const params = {
+    'user_id': 500,
+    'User-Agent': 'MyApp/1.0',
+};
+
+const params2 = {
+    user_id: 50,
+};
+
+void params;
+void params2;