+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to my ESLint config will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### Fixed
+- loosened `@typescript-eslint/naming-convention` and `no-underscore-dangle` rules in certain cases.
+
 ## [6.2.0] - 2022-02-22
 ### Added
 - Support for eslint-plugin-sonarjs@0.12